Rename files in multibody to follow cppguide
