+++ drake/multibody/constraint/test/rigid_body_constraint_test.cc
